merge changes (#1)
* icon/link space

* Fixed text disappearing in mobile version navbar

* Fix attribution

Following a discussion on
https://github.com/sugarlabs/www-sugarlabs/pull/192#issuecomment-429739009
the consensus seemed to be to change this wording.

* Centered pagination for carousel

the buttons for the middle portion (owl carousel under "A GLIMPSE OF SUGAR..." on index.html were left aligned, so this commit changes that to be centered.

* Remove minified airspace.css

Site is 4.96 MB over 79 requests, and the minified airspace.css saves
only a few network packets among thousands.  The minified file hinders
review of pull requests.  Remove it and fix references to it.

* Remove minified main.js and plugins.js

No need for them to be minified.

* Resequence lists

* Newsletter subscribe button - change spacing and width

Fixes https://github.com/sugarlabs/www-sugarlabs/issues/181

* changed profiles.md to profiles.html

Changed the look and feel of profile.md file using the html tag and changed the file type to .html. The format is as per the task assinged - (3 users per row in a table with cell spacing and padding.

* Add second Google verification file

To give samson goddy access to Google knownledge graph

* Update airspace.css

* Update nav.html

* Update airspace.css

* Update nav.html

* Update airspace.css

* Update nav.html

* Update airspace.css

* Update airspace.css

* Update nav.html

* Update airspace.css

* Update airspace.css

* Update nav.html

* Update airspace.css

* Revert "Webpage Styling for Google-Code in"

* Remove Sugar Labs Embedded Youtube Video

* Remove unused syntax.css

* Remove HTML5shiv from updated modernizr

* Replace redundant Social Share Kit with Ion Icons

* Remove unused .sass files

* Remove duplication between README.md and documentation.md (#247)

* Remove Chat

- it doesn't work at the moment, leading people astray,

- it isn't clear how to make it work.

Work in progress; there may be other things to remove or change.

* Update index.html

* Update index.html